requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
Django==3.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-rest-action-permissions==2.0.0
django-utils-six==2.0
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.1.0
PyJWT==1.7.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
